requirements.txt
celery[redis]==5.2.7
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary
